requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.110.0
flatbuffers==23.5.26
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
joblib==1.3.2
keras==3.0.5
numpy==1.26.4
openai==1.12.0
pandas==2.2.1
protobuf==4.25.3
pydantic==2.6.2
requests==2.31.0
rich==13.7.0
scikit-learn==1.4.1.post1
scipy==1.12.0
tensorflow==2.16.0rc0
tqdm==4.66.2
typing_extensions==4.10.0
uvicorn==0.27.1
wrapt==1.16.0
yfinance==0.2.37
slowapi==0.1.8
TA-Lib==0.4.28
newsapi-python==0.2.7
transformers==4.37.2
